+++ PyWebRunner/WebRunner.py
@@ -11,7 +11,7 @@ import pkg_resources
 import re
 import yaml
 import json
-from PyWebRunner.utils import which, Timeout, fix_firefox, fix_chrome, prompt
+from PyWebRunner.utils import which, Timeout, fix_firefox, fix_chrome, prompt, download_file
 from xvfbwrapper import Xvfb
 from selenium import webdriver
 from selenium.common.exceptions import (NoSuchElementException, NoSuchWindowException,
@@ -718,6 +718,40 @@ class WebRunner(object):
         sel = Select(elem)
         sel.select_by_value(value)
 
+
+    def download(self, url, filepath):
+        '''
+        Download a file.
+
+        Parameters
+        ----------
+        filepath: str
+            The location and name of the file.
+        src: str
+            The URL to download
+        '''
+        download_file(url, filepath)
+
+
+    def save_image(self, filepath, selector=None, elem=None):
+        '''
+        Download an image.
+
+        Parameters
+        ----------
+        filepath: str
+            The location and name of the file.
+        selector: str
+            Any valid CSS selector
+        '''
+        if selector:
+            elem = self.get_element(selector)
+
+        if elem:
+            src = elem.get_attribute('src')
+            download_file(src, filepath)
+
+
     def move_to(self, selector, click=False):
         '''
         Move to the element matched by selector or passed as argument.
@@ -796,7 +830,7 @@ class WebRunner(object):
         elem = self.find_element(selector)
         return elem
 
-    def get_text(self, selector):
+    def get_text(self, selector=None, elem=None):
         '''
         Gets text from inside of an element by CSS selector.
 
@@ -811,7 +845,9 @@ class WebRunner(object):
             The text from inside of a selenium element object.
 
         '''
-        elem = self.get_element(selector)
+        if not elem:
+            elem = self.get_element(selector)
+
         return str(elem.text)
 
     def get_texts(self, selector):